IOS-3014 add web view module
