--- Tangem/Modules/SprinklrSupportChat/SprinklrSupportChatRoutable.swift
@@ -1,13 +1,0 @@
-//
-//  SprinklrSupportChatRoutable.swift
-//  Tangem
-//
-//  Created by Sergey Balashov on 15.02.2023.
-//  Copyright © 2023 Tangem AG. All rights reserved.
-//
-
-import Foundation
-
-protocol SprinklrSupportChatRoutable: AnyObject {
-
-}
